Remove templateAgent: move Agent implementations into claude/ and pi/ packages
- claude.Agent and pi.Agent are now concrete types with Run, RunInteractive, BuildPrompt
- Extracted monitorAgent into agents/runner subpackage to avoid import cycles
- NewAgent factory returns claude.New() or pi.New() directly
- Deleted template_agent.go (replaced by per-package implementations)
- Updated tests to use concrete agent types instead of *templateAgent
